+++ rust/guest/k16-kernel/src/main.rs
@@ -3,11 +3,17 @@
 
 extern crate k16_rt;
 
+mod console;
 mod control;
 mod debug;
+mod font;
+mod generated;
+mod gpu;
 mod init;
 mod mmio;
 mod syscall;
+mod terminal;
+mod terminal_render;
 mod timer;
 mod trap;
 
@@ -15,6 +21,7 @@ use core::panic::PanicInfo;
 
 #[no_mangle]
 pub extern "C" fn _start() -> ! {
+    console::init();
     debug::print_kernel_ok();
     trap::initialize();
     control::set_ready();